backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0
httpx>=0.26.0

# News API
newsapi-python>=0.2.7

# Google Earth Engine
earthengine-api>=0.1.390
google-auth>=2.27.0

# Sentiment Analysis
textblob>=0.18.0
google-generativeai>=0.3.2

# Financial Data
polygon-api-client>=1.13.4

# Caching
cachetools>=5.3.2

# Geospatial Processing
numpy>=1.24.0
Pillow>=10.0.0
